Fix typecheck-failing functions not reaching declare()
When one of the body statements of a function fails typecheck, the whole
typecheck() early-exits, preventing the function to be declared in the
scope and thus causing a lot of dependent errors on the code that uses
the function.  This will be inconvenient as changing a body of a
function would cause a lot of previously unseen errors in different
places that the user did not expect to show up.  We therefore allow a
function to fail typecheck for its body statements, but still be
declare()-ed in its scope.