parse: separate typeexpr from type
It seems better to have an additional AST node kind for types separated
from the actual type object, because otherwise it conflates parsing code
with semantic code in terms of data structure.  We have to determine if
a type object is fully typechecked yet or not, and especially for value
("canonical") types, make sure that the *type pointer points to the
original declaration and not just the AST node.