Use linked list for storing array of node
The rationale is to make it easier to rewrite the AST.  Linked list is
much easier than arrays to add/delete an element in the list *when
visiting the element*, because each element maintains the next pointer
themselves.

One peeve is that being a single linked list, it's only possible to add
right next to a target node, so we have to keep track of the last node
in the list for appending.  Maybe making this into a separate struct
might be better.